(PUP-1253) Add systemd service masking support
This updates the systemd service type to add support for masking. If a
service is masked, it is deemed to also be disabled. If a service is
masked and changed to enabled, it will first be unmasked since the
standard 'systemctl enable' command does not properly unmask the command
first.

The check against the delcared feature is in place so that users do not
have update their code due to differences between systemd and
non-systemd systems.

Conflicts:
	spec/unit/provider/service/systemd_spec.rb