(PUP-4503)(PUP-2023) Update service provider default for newer debian/ubuntu releases
With the release of Debian 8 (Jessie) and Ubuntu 15.04 (Vivid Vervet),
ubuntu and debian now default to systemd as the main service manager,
rather than init or upstart. This commit updates Puppet to
use systemd as the default service provider on these two systems, while
also defaulting back to other service providers if necessary.

Additionally, there is a bug on debian systemd where it is unable to
correctly detect SysV style services[1]. This commit also adds in a
workaround to correctly detect `is-enabled` for these services.

[1] - https://bugs.debian.org/cgi-bin/bugreport.cgi?bug=751638

Conflicts:
	lib/puppet/provider/service/debian.rb
	lib/puppet/provider/service/systemd.rb
	spec/unit/provider/service/debian_spec.rb
	spec/unit/provider/service/systemd_spec.rb
	spec/unit/provider/service/upstart_spec.rb